requirements.txt
selenium~=4.1.3
pytest~=7.1.2
allure-pytest~=2.9.45
webdriver-manager~=3.5.4
requests~=2.27.1
PyYaml~=6.0